requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
python-dotenv==0.19.0
Flask-WTF==1.0.0
SQLAlchemy==1.4.48
WTForms==3.0.1